requirements.txt
agora_token_builder==1.0.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
cloudinary==1.41.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.2.0
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-htmx==1.21.0
django-redis==5.4.0
django-tailwind==3.8.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==24.2.1
python-decouple==3.8
python-magic==0.4.27
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.5.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tailwind==3.1.5b0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.1.1
